Refactor delete saved button/functionality

+++ src/scripts.js
@@ -44,10 +44,14 @@ navViewProfileButton.addEventListener( 'click' , showCookingProfile);
 recipeCardGridContainer.addEventListener('click', ( e ) => {
     if (e.target.classList == 'save-button') {
       return  saveRecipeToRecipesToCook( e );
-    } else { (e.target.classList == 'remove-button') 
-}
-return deleteRecipeFromRecipesToCook( e );
-})
+    };
+}) 
+
+recipeCardGridContainer.addEventListener('click', ( e ) => {
+if(e.target.classList == 'remove-button') {
+ return deleteRecipeFromRecipesToCook( e );
+    };
+});
 
 recipeCardGridContainer.addEventListener('click', (e) => {
     if (e.target.classList == 'lets-make-it-button') {
@@ -167,7 +171,7 @@ function displayRecipeInfo( e ){
         if( e.target.id == dish.id ){
             recipeClass = new Recipe( dish, dish.ingredients );
             recipeClass.getIngredientsWithNames(dish.ingredients, ingredientList);
-            console.log(dish);
+            console.log('DEEEZ:', dish);
             h4.innerText = dish.name;  
             instructionText.innerText = dish.instructions.map( task => `${ task.number }: ${ task.instruction }` ).join( '  ' );
             ingredientText.innerText = dish.ingredients.map( foodItem => ` ${ ( foodItem.quantity.amount ).toFixed( 2 ) } ${ foodItem.quantity.unit } ${ foodItem.name }` );
@@ -227,9 +231,20 @@ function deleteRecipeFromRecipesToCook( e ) {
     return newRecipe.recipes.find((removeDish) => {
     if(e.target.id == removeDish.id) {
        currentUser.removeRecipeFromRecipesToCook(removeDish)
+
+    //    console.log('FUNCTION', displayRecipeInfo( e ))
+    //    displayRecipeInfo( e );
     }
+    //forEach targeted id, remove from HTML?
     //ERROR READING .includes in this function -- to fix
     // console.log('REMOVED RECIPES: ', currentUser.recipesToCook)
+    console.log('ARRAY: ', currentUser.recipesToCook)
     return currentUser.recipesToCook
+    // return recipeCard.innerHTML = `this has been removed`
+   //^^change the innerHTML to reflect it being removed
     })
 }
+
+// function removeRecipeCardFromDisplay () {
+//     return currentUser.recipesToCook
+// }
